Import answers to exercises for ch1 & ch2
They are previously developed as run in scala REPL and kept in the notes
repo. We now decided to import them here as
src/main/scala/org/learningconcurrency/answers/chX/chX.scala. The
original code is kept as is, each enclosed within a dummy object in
order to compile successfully. They are still intended to run under REPL
as there are endless loops.